requirements.txt
slack-bolt==1.23.0
slack-sdk==3.35.0
python-dotenv==1.1.0
faiss-cpu==1.11.0
sentence-transformers==4.1.0
transformers==4.52.1
scikit-learn==1.6.1
torch==2.7.0
watchdog==6.0.0
rapidfuzz==3.13.0
numpy==2.2.6
huggingface-hub==0.31.4
requests==2.32.3
tqdm==4.67.1
huggingface_hub[hf_xet]
